backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
playwright==1.47.0
google-genai>=1.42.0
python-dotenv==1.0.1
pillow==10.4.0
